+++ src/main/java/org/olat/group/model/BusinessGroupViewImpl.java
@@ -50,32 +50,32 @@ public class BusinessGroupViewImpl extends PersistentObject implements BusinessG
 	private Boolean autoCloseRanksEnabled;
 	private Date lastModified;
 
-	private int numOfOwners;
-	private int numOfParticipants;
-	private int numOfRelations;
-	private int numOfOffers;
-	private int numOfValidOffers;
+	private long numOfOwners;
+	private long numOfParticipants;
+	private long numOfRelations;
+	private long numOfOffers;
+	private long numOfValidOffers;
 
 	@Override
-	public int getNumOfRelations() {
+	public long getNumOfRelations() {
 		return numOfRelations;
 	}
 
-	public void setNumOfRelations(int numOfRelations) {
+	public void setNumOfRelations(long numOfRelations) {
 		this.numOfRelations = numOfRelations;
 	}
 
 	@Override
-	public int getNumOfOwners() {
+	public long getNumOfOwners() {
 		return numOfOwners;
 	}
 
-	public void setNumOfOwners(int numOfOwners) {
+	public void setNumOfOwners(long numOfOwners) {
 		this.numOfOwners = numOfOwners;
 	}
 
 	@Override
-	public int getNumOfParticipants() {
+	public long getNumOfParticipants() {
 		return numOfParticipants;
 	}
 
@@ -83,22 +83,21 @@ public class BusinessGroupViewImpl extends PersistentObject implements BusinessG
 		this.numOfParticipants = numOfParticipants;
 	}
 
-
 	@Override
-	public int getNumOfOffers() {
+	public long getNumOfOffers() {
 		return numOfOffers;
 	}
 	
-	public void setNumOfOffers(int numOfOffers) {
+	public void setNumOfOffers(long numOfOffers) {
 		this.numOfOffers = numOfOffers;
 	}
 
 	@Override
-	public int getNumOfValidOffers() {
+	public long getNumOfValidOffers() {
 		return numOfValidOffers;
 	}
 
-	public void setNumOfValidOffers(int numOfValidOffers) {
+	public void setNumOfValidOffers(long numOfValidOffers) {
 		this.numOfValidOffers = numOfValidOffers;
 	}
